refactor: Extracted SeedDataProvider.cs and InMemoryDatabaseProvider.cs into a separate project
